changemtn necessire pour sgbd sqlserver

+++ database/migrations/2025_09_23_103739_add_definition_to_b_rec_ticket_table.php
@@ -12,7 +12,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('b_rec_tickets', function (Blueprint $table) {
-            $table->text('definition')->nullable();
+             if (!Schema::hasColumn('b_rec_tickets', 'definition')) {
+                $table->text('definition')->nullable();
+            }
         });
     }
 
@@ -22,7 +24,9 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('b_rec_tickets', function (Blueprint $table) {
-            $table->dropColumn('definition');
+            if (Schema::hasColumn('b_rec_tickets', 'definition')) {
+                $table->dropColumn('definition');
+            }
         });
     }
 };
